+++ examples/egg/config/plugin.js
@@ -1,3 +1,4 @@
 'use strict'
 
 exports.router = true
+exports.cors = { enable: true, package: 'egg-cors' }